fix_some_bag

+++ tasks/olesnitskiy_v_find_viol/tests/performance/main.cpp
@@ -11,17 +11,17 @@ class OlesnitskiyVFindViolPerfTests : public ppc::util::BaseRunPerfTests<InType,
   InType input_data_;
 
   void SetUp() override {
-    const size_t vector_size = 90000000;
+    const int vector_size = 90000000;
     std::vector<double> vector(vector_size);
 
-    for (size_t i = 0; i < vector_size; i++) {
+    for (int i = 0; i < vector_size; i++) {
       vector[i] = static_cast<double>(i % 1000) + ((i & 1) == 0 ? 1.0 : -1.0);
     }
     input_data_ = vector;
   }
 
   bool CheckTestOutputData(OutType &output_data) final {
-    return output_data >= 0 && static_cast<size_t>(output_data) <= input_data_.size();
+    return output_data >= 0 && static_cast<int>(output_data) <= static_cast<int>(input_data_.size());
   }
 
   InType GetTestInputData() final {
